+++ plugins/amazonq/codewhisperer/jetbrains-community/src/software/aws/toolkits/jetbrains/services/codewhisperer/util/CodeWhispererEndpointCustomizer.kt
@@ -26,7 +26,7 @@ import software.amazon.awssdk.services.cognitoidentity.CognitoIdentityClient
 import software.aws.toolkits.core.ToolkitClientCustomizer
 import software.aws.toolkits.jetbrains.core.AwsSdkClient
 import software.aws.toolkits.jetbrains.services.codewhisperer.explorer.CodeWhispererExplorerActionManager
-import software.aws.toolkits.jetbrains.services.codewhisperer.settings.CodeWhispererSettings
+import software.aws.toolkits.jetbrains.settings.CodeWhispererSettings
 import software.aws.toolkits.jetbrains.services.telemetry.AwsCognitoCredentialsProvider
 import java.net.Proxy
 import java.net.URI
